Move parser now returns multiple interpretations
There is at least one case where a move could mean more than one
thing. See "bc4". This move could either mean "pawn on file B takes on
c4", or "Bishop to c4". The parser has no way of knowing which one the
user meant without knowing about the state of the board. The solution to
this is that now the parser returns multiple possible interpretations of
the move, and the rest of the program is responsible of validating and
processing each one and execute the right one. Note: "Bc4" can only be
interpreted as "bishop to c4" because of the capitalization.